add type selection for direction, add select default home/work feature

+++ controllers/location-controller.ts
@@ -109,7 +109,7 @@ async function updateLocation(req: Request, res: Response) {
     const { locId } = req.params;
 
     // Validating request
-    const { name, street, city, province } = req.body
+    const { name, street, city, province, isHome, isWork } = req.body
 
     if (!name || !street || !city || !province) {
         return res.status(400).json({ message: "Must provide location name and address" })
@@ -126,8 +126,56 @@ async function updateLocation(req: Request, res: Response) {
     });
 
     // If name is already taken
-    if (isNameExist) {
+    if (isNameExist && String(isNameExist.id) !== locId) {
         return res.status(400).json({ message: "Name Already in Use: Please select a different name" })
+    }
+
+    // Retrieve current default
+    const defaultIds = await getDefault(profile_id);
+    const defaultAddress = `${street} ${city} ${province}`.replaceAll(' ', '+');
+
+    // If isHome is true 
+    if (isHome) {
+        // If it was defaultWork don't allow
+        if (Number(locId) === defaultIds[1]) {
+            return res.status(400).json({ message: "Default work must always be present." })
+        }
+        // If it was not originally defaultHome
+        if (Number(locId) !== defaultIds[0]) {
+            // Update existing default to false
+            await prisma.location.update({
+                where: { id: defaultIds[0] },
+                data: { isHome: false }
+            })
+
+            // Update profile 
+            await prisma.user_Profile.update({
+                where: { id: profile_id },
+                data: { default_home: defaultAddress }
+            })
+        }
+    }
+
+    // If isWork is true 
+    if (isWork) {
+        // If it was defaultHome don't allow
+        if (Number(locId) === defaultIds[0]) {
+            return res.status(400).json({ message: "Default home must always be present." })
+        }
+        // If it was not originally defaultWork
+        if (Number(locId) !== defaultIds[1]) {
+            // Update existing default to false
+            await prisma.location.update({
+                where: { id: defaultIds[1] },
+                data: { isHome: false }
+            })
+
+            // Update profile 
+            await prisma.user_Profile.update({
+                where: { id: profile_id },
+                data: { default_work: defaultAddress }
+            })
+        }
     }
 
     // Updating the data
@@ -137,7 +185,7 @@ async function updateLocation(req: Request, res: Response) {
             user_ProfileId: profile_id
         },
         data: {
-            name, street, city, province
+            name, street, city, province, isHome, isWork
         }
     });
 
@@ -159,6 +207,13 @@ async function deleteLocation(req: Request, res: Response) {
         return res.status(400).json({ message: `Unable to find your location with ID: ${locId}` });
     };
 
+    // If it's default location
+    const [homeId, workId] = await getDefault(profile_id);
+
+    if (Number(locId) === homeId || Number(locId) === workId) {
+        return res.status(400).json({message: "You cannot delete default location."})
+    }
+
     // Delete information
     await prisma.location.delete({
         where: {
@@ -168,6 +223,25 @@ async function deleteLocation(req: Request, res: Response) {
     });
 
     return res.sendStatus(204);
+}
+
+// Check if user has default 
+async function getDefault(profile_id) {
+    const home = await prisma.location.findFirst({
+        where: {
+            user_ProfileId: profile_id,
+            isHome: { not: false },
+        }
+    })
+
+    const work = await prisma.location.findFirst({
+        where: {
+            user_ProfileId: profile_id,
+            isWork: { not: false },
+        }
+    })
+
+    return [home.id, work.id];
 }
